requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-dotenv
pandas
redis
aiofiles
requests
asyncpg
apscheduler
httpx
